fix rerendering transaction component get all data in explorer
